Changed location dialog to show distances to autocomplete suggestions. When locations added, sets correct latLng depending on what was selected

+++ src/main/java/com/chdryra/android/reviewer/VhdLocatedPlaceDistance.java
@@ -40,8 +40,8 @@ public class VhdLocatedPlaceDistance implements ViewHolderData {
         return mPlace;
     }
 
-    public int getDistance() {
-        return (int) mDistanceAndBearings[0];
+    public float getDistance() {
+        return mDistanceAndBearings[0];
     }
 
     @Override
